requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
psycopg==3.2.5
gunicorn==23.0.0
whitenoise==6.8.2
django-redis==5.4.0
redis==5.2.1
dj-database-url==2.3.0
django-cors-headers==4.6.0
